refactor(routes): refactor route to prepare api
