+++ src/server/bootstrap/configureDatabase.js
@@ -1,7 +1,7 @@
 import Knex from 'knex';
 import { Model } from 'objection';
 import config from 'config';
-import knexConfig from '../../knexfile';
+import knexConfig from '../../../knexfile';
 
 export default () => {
   const knex = Knex(knexConfig[config.get('env')]);
