+++ src/server/bootstrap/setup.js
@@ -1,0 +1,5 @@
+import configureDatabase from 'server/bootstrap/configureDatabase';
+import handleKillSignals from 'server/bootstrap/handleKillSignals';
+
+handleKillSignals();
+configureDatabase();
