+++ src/server/middlewares/csp.js
